split dynlib and packet_processor
