+++ src/bgp/bgp.c
@@ -27,7 +27,7 @@
 #include "bgp_blackhole.h"
 #include "bgp_ls.h"
 #include "thread_pool.h"
-#include "custom_packet_parsing/packet_processor.h"
+#include "packet_processor/packet_processor.h"
 #if defined WITH_EBPF
 #include "ebpf/ebpf_rp_balancer.h"
 #endif
